% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctmref-package.R
\docType{package}
\name{mctmref-package}
\alias{mctmref-package}
\alias{mctmref}
\title{mctmref: Multivariate Conditional Reference Regions via Transformation Models}
\description{
Estimates bivariate conditional reference and tolerance regions based on
Multivariate Conditional Transformation Models (MCTMs).
}
\section{Main functions}{

\describe{
  \item{\code{\link{mctm_fit}}}{Fit a bivariate MCTM.}
  \item{\code{\link{mmlt_perc}}}{Marginal conditional percentile curves.}
  \item{\code{\link{bivreg_mctm}}}{Bivariate conditional reference region.}
  \item{\code{\link{tol_mctm}}}{Bivariate conditional tolerance region.}
  \item{\code{\link{mctm_zscores}}}{Joint Z scores and decorrelated residuals.}
  \item{\code{\link{mctm_mahalanobis}}}{Mahalanobis distance in Z space.}
  \item{\code{\link{mctm_loglik}}}{Joint log-likelihood per observation.}
  \item{\code{\link{mctm_loglik_grid}}}{Joint density on a response grid.}
  \item{\code{\link{mctm_loglik_by_age}}}{Log-likelihood profile over the covariate.}
}
}

\references{
Lado-Baleato, O., Cadarso-Suarez, C., Kneib, T. and Gude, F. (2023).
Multivariate reference and tolerance regions based on conditional
transformation models. \emph{Biometrical Journal}, 65, 2200229.
\doi{10.1002/bimj.202200229}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bioscar/mctmref}
  \item Report bugs at \url{https://github.com/Bioscar/mctmref/issues}
}

}
\author{
\strong{Maintainer}: Oscar Lado-Baleato \email{ladooscar@uniovi.es}

Authors:
\itemize{
  \item Thomas Kneib \email{tkneib@uni-goettingen.de}
}

}
